# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flapjack Cloud API",
    "description": "Cloud infrastructure and billing platform API",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/billing/estimate": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "`GET /billing/estimate` — compute an estimated invoice for a given month.",
        "description": "**Auth:** JWT (`AuthenticatedTenant`).\nAccepts an optional `?month=YYYY-MM` query param (defaults to current month).\nRuns the invoice computation pipeline against live usage data without\npersisting the result. The response includes line items, subtotal/total,\nand whether the plan minimum was applied.",
        "operationId": "get_estimate",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "description": "Billing month in YYYY-MM format (defaults to current month).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estimated bill for the month",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstimatedBillResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/billing/payment-methods": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "`GET /billing/payment-methods` — list saved payment methods from Stripe.",
        "description": "**Auth:** JWT (`AuthenticatedTenant`).\nResolves the customer's `stripe_customer_id` and fetches all attached\npayment methods from Stripe. Returns 400 if no Stripe customer is linked.",
        "operationId": "list_payment_methods",
        "responses": {
          "200": {
            "description": "List of payment methods",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentMethodResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/billing/payment-methods/{pm_id}": {
      "delete": {
        "tags": [
          "Billing"
        ],
        "operationId": "delete_payment_method",
        "parameters": [
          {
            "name": "pm_id",
            "in": "path",
            "description": "Payment method identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Payment method deleted"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Payment method not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/billing/payment-methods/{pm_id}/default": {
      "post": {
        "tags": [
          "Billing"
        ],
        "operationId": "set_default_payment_method",
        "parameters": [
          {
            "name": "pm_id",
            "in": "path",
            "description": "Payment method identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Default payment method set"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Payment method not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/billing/portal": {
      "post": {
        "tags": [
          "Billing"
        ],
        "summary": "Create a Stripe billing portal session for the authenticated customer.",
        "operationId": "create_billing_portal_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBillingPortalSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Billing portal session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingPortalSessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Stripe not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/billing/publishable-key": {
      "get": {
        "tags": [
          "Billing"
        ],
        "operationId": "get_publishable_key",
        "responses": {
          "200": {
            "description": "Stripe publishable key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishableKeyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Stripe publishable key unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateBillingPortalSessionRequest": {
        "type": "object",
        "required": [
          "return_url"
        ],
        "properties": {
          "return_url": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Shared error envelope returned by all API error responses.\nShape: `{\"error\": \"<message>\"}`.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error description."
          }
        }
      },
...
      "EstimatedBillResponse": {
        "type": "object",
        "required": [
          "month",
          "subtotal_cents",
          "total_cents",
          "line_items",
          "minimum_applied"
        ],
        "properties": {
          "line_items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EstimateLineItem"
            }
          },
          "minimum_applied": {
            "type": "boolean"
          },
          "month": {
            "type": "string"
          },
          "subtotal_cents": {
            "type": "integer",
            "format": "int64"
          },
          "total_cents": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PaymentMethodResponse": {
        "type": "object",
        "required": [
          "id",
          "card_brand",
          "last4",
          "exp_month",
          "exp_year",
          "is_default"
        ],
        "properties": {
          "card_brand": {
            "type": "string"
          },
          "exp_month": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "exp_year": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "id": {
            "type": "string"
          },
          "is_default": {
            "type": "boolean"
          },
          "last4": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearer_jwt": []
    }
  ]
}